Refactor : Rename Package to domain from entity
엔티티라는 페키지 네임보다 도메인이 더 적합하다 생각이 되어 변경
