+++ employee/EmployeeManager.java
@@ -1,9 +1,9 @@
 package employee;
 
-import employee.entity.Employee;
-import employee.entity.Manager;
-import employee.entity.Secretary;
-import employee.entity.Staff;
+import employee.domain.Employee;
+import employee.domain.Manager;
+import employee.domain.Secretary;
+import employee.domain.Staff;
 
 import java.util.ArrayList;
