+++ employee/domain/RestrictionLevel.java
@@ -1,4 +1,4 @@
-package employee.entity;
+package employee.domain;
 
 import lombok.AccessLevel;
 import lombok.Builder;
